requirements.txt
jsonschema==2.3.0
validators==0.14.2
filetype==1.0.5
insightconnect-integrations-plugin-spec-tooling==1.3.1
requests==2.22.0
urlextract==0.14.0
typing==3.7.4.1
PyYAML==5.4.1
dacite==1.6.0
pydantic==1.8.2
